java-mid2/src/main/java/org/example/chap6_collection_hash/MyHashSetV0.java
package org.example.chap6_collection_hash;

import java.util.Arrays;

public class MyHashSetV0 {

    private int[] elementData = new int[10];
    private int size = 0;

    public boolean add(final int value) {
        if (contains(value)) {
            return false;
        }
        elementData[size++] = value;
        return true;
    }

    public boolean contains(final int value) {
        for (int data : elementData) {
            if (data == value) {
                return true;
            }
        }
        return false;
    }

    public int size() {
        return size;
    }

    @Override
    public String toString() {
        return "MyHashSetV0{" +
                "elementData=" + Arrays.toString(Arrays.copyOf(elementData, size)) +
                ", size=" + size +
                '}';
    }

}
